refactor(blast): Use null to represent destroyed cells in the grid
Updated the applyBlastToGrid function to change its representation of
destroyed cells. Previously, affected cells had their oreType property
set to the string "destroyed". This commit changes that behavior to set
the entire cell's data to null. Using null is a more semantically
correct and idiomatic way to represent an empty or non-existent slot in
the grid data structure. This refactoring of the data layer paves the
way for the rendering component to treat these empty slots as hollow,
rather than as a special type of block.